.history/requirements_20250304144943.txt
requests==2.31.0
flask==2.3.3
gunicorn>=22.0.0
python-dotenv==1.0.0
sqlalchemy==2.0.20
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.2.2
psycopg2-binary==2.9.7